requirements.txt
asyncio
discord.py
psutil
